% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async_get_sreality_detail.R
\name{async_get_sreality_detail}
\alias{async_get_sreality_detail}
\title{Parsing function for asynchronous scraping of listing details}
\usage{
async_get_sreality_detail(sreality_detail_url)
}
\arguments{
\item{sreality_detail_url}{The url to parse}
}
\value{
CRRRI Promise
}
\description{
This function is to be used exclusively with the crrri::perform_with_chrome function
}
