% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sreality_scraper.R
\name{get_page_info}
\alias{get_page_info}
\title{Parse scraped listings lists (paginations) into a dataframe}
\usage{
get_page_info(html_dump_path)
}
\arguments{
\item{html_dump_path}{The path to a respective html dump of a listings page}
}
\value{
A tibble with extracted url, name, location and price
}
\description{
Parse scraped listings lists (paginations) into a dataframe
}
\examples{
\dontrun{get_page_info("html_dump/page.html")}
}
