% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sreality_scraper.R
\name{get_url_variants}
\alias{get_url_variants}
\title{Returns pagination links from Sreality query}
\usage{
get_url_variants(sreality_query)
}
\arguments{
\item{sreality_query}{the url of a search query on sreality.cz}
}
\value{
a character vector of urls to scrape
}
\description{
Returns pagination links from Sreality query
}
\examples{
\dontrun{get_url_variants("https://www.sreality.cz/hledani/prodej/byty")}
}
