% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sreality_detail.R
\name{parse_flat_detail}
\alias{parse_flat_detail}
\title{Parse a listing_detail html dump into a dataframe}
\usage{
parse_flat_detail(html_detail_path)
}
\arguments{
\item{html_detail_path}{the path to the listing detail html file}
}
\value{
a tibble with the scraped attibutes
}
\description{
This function takes the path to a dumped html file with a listing details and extracts most of the info there.
The param fields are downloades as a nested list in a single column to prevent data corruption on download.
}
\examples{
#' \dontrun{parse_flat_detail("data/sreality_listing_detail.html")}

}
