% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scrape_functions.R
\name{save_as_html}
\alias{save_as_html}
\title{Dump HTML files of a list of webpages}
\usage{
save_as_html(..., output_dir = ".")
}
\arguments{
\item{...}{a character list of urls to scrape}

\item{output_dir}{the output directory folder where to dump the html files}
}
\value{
a directory of scraped html files
}
\description{
uses a headless chrome browser to render a list of urls and dump them as html files to a selected folder
}
\examples{
\dontrun{save_as_html("https://test.com", "https://google.com",output_dir = "dump")}
}
