% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sreality.R
\name{scrape_sreality}
\alias{scrape_sreality}
\title{Title}
\usage{
scrape_sreality(
  query = "https://www.sreality.cz/hledani/prodej/byty?vlastnictvi=osobni&bez-aukce=1",
  tag = stringr::str_extract(query,
    "(?<=\\\\/)[^\\\\/]+(?=\\\\?)|(?<=\\\\/)[^\\\\/]+$")
)
}
\arguments{
\item{query}{the query of the search on sreality defaults to https://www.sreality.cz/hledani/prodej/byty/}

\item{tag}{(char) optional - will include a tag into the data if used}
}
\value{
Creates a csv file with the parsed listings and a log file in the specified folder
}
\description{
Title
}
\examples{
\dontrun{scrape_sreality("https://www.sreality.cz/hledani/prodej/byty")}
}
