% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scrape_functions.R
\name{async_save_as_html}
\alias{async_save_as_html}
\title{Asynchronously save an webpage as a HTML Document}
\usage{
async_save_as_html(url, folder)
}
\arguments{
\item{url}{the url to scrape}

\item{folder}{the target folder to dump the html files}
}
\value{
a promise of the scraped page
}
\description{
This is a working function to asynchronously receive DOM objects
}
