% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sreality_pagination.R
\name{get_page_info}
\alias{get_page_info}
\title{Returns the extracted data from the html dumps from the listings page}
\usage{
get_page_info(html_dump_path)
}
\arguments{
\item{html_dump_path}{The path to a respective html dump of a listings page}
}
\value{
A tibble with extracted url, name, location and price
}
\description{
Returns the extracted data from the html dumps from the listings page
}
\examples{
\dontrun{get_page_info("html_dump/page.html")}
}
