% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sreality_pagination.R
\name{get_sreality_listings}
\alias{get_sreality_listings}
\title{Get the number of listing on a specific Sreality.cz search query}
\usage{
get_sreality_listings(sreality_query)
}
\arguments{
\item{sreality_query}{the url of a search query on sreality.cz}
}
\value{
the number of listings to be found on the site
}
\description{
Get the number of listing on a specific Sreality.cz search query
}
\examples{
\dontrun{get_sreality_listings("https://www.sreality.cz/hledani/prodej/byty")}
}
